+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## 1.12.0 (Unreleased)
 
 FEATURES:
+* **New Data Source**: `tencentcloud_clb_instances`
 * **New Data Source**: `  tencentcloud_dc_instances`
 * **New Data Source**: `tencentcloud_dcx_instances`
+* **New Resource**: `tencentcloud_clb_instance`
 * **New Resource**: `tencentcloud_dcx`
 
 BUG FIXIES: